requirements.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dj-database-url==2.0.0
psycopg2==2.9.6
python-decouple==3.8
Pillow==9.5.0
gunicorn==20.1.0

coverage==7.2.5
django-fake-model==0.1.4
django-filter==23.2
django-phonenumber-field[phonenumbers]
django-storages[azure]==1.12.3
freezegun==1.2.2
model-bakery==1.11.0
parameterized==0.9.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-subtests==0.10.0
pytest==7.3.1
